requirements.txt
django-crispy-forms
django-countries
django-filter
mtgsdk
braintree